WIP: freedreno: add CS_PVT_MEM regs and the stubs using them
The first step to add private memory support for a5xx.

This adds the registers definitions (assuming that they are the
same as in a6xx, just the base address is different), and adds
the code to fd5_emit_shader which does write to these registers,
but uses the values (zeroes) which correspond to private memory
size of zero. A quick sanity test didn't show a breakage to the
currently working CTS basic tests.

Signed-off-by: Andrey Konovalov <andrey.konovalov@linaro.org>