WIP: freedreno: set up private memory for a530
This is the second step to add private memory support for a5xx. It
is currently enabled for a530 only, as the other a5xx parts weren't
tested.

* Make the number of SP cores and the number of fibers per SP to be set
  to non-zero values in the struct fd_context. The values themselves
  are yet to be confirmed.
* Calculate private memory size per fiber and per SP using the data
  from fd_context, and pvtmem_size from ir3_shader_variant (pvtmem_size
  is set correctly for a5xx).
* Program private memory size per item, total private memory size, and
  private memory hardware stack offset values into the GPU registers.

This code is written in the assumption that a5xx isn't different from
a6xx as regards to the scope of this patch. This needs to be verified,
and some corrections are likely to be needed. Still the current code
made the below test cases from the OpenCL CTS basic tests to pass OK:
    * vload_* and vstore_* - except for the *_local ones
    * vector_creation

Signed-off-by: Andrey Konovalov <andrey.konovalov@linaro.org>